Crate fdb_rl

FoundationDB Record Layer

This crate is work-in-progress.

At this time, the thinking with Record Layer is that it will consist of a schema (RecordSchema) containing multiple RecordEntitys.

Each RecordEntity will map to seralized byte array such as a Protobuf Message and will have a primary key tuple and optionally one or more secondary indexes.

This crate will provide:

APIs for Schema and Index management.

Ensure transactionally consistent secondary indexes during inserts, updates and deletes.

A very simple API for querying based on primary key and secondary indexes.

In addition, a simple transactionally consistent queuing system similar to QuiCK will also be implemented.

There are no plans to implement a query planner or other higher level database features.

Modules

cursor

Provides types and traits for Record Layer cursors.

error

Provides error constants for use with FdbError and FdbResult.

range

Provides types for Record Layer range.

raw_record 🔒

Provides RawRecord type and associated items.

record_context 🔒

Provides RecordContext type and associated items.

record_version 🔒

Provides RecordVersion type and associated items.

scan

Provides types for out-of-band scan limits in Record Layer.

split_helper 🔒

Helper module for splitting records across multiple key-value pairs.

test

Types and Functions meant only for integration testing.

Structs

RecordContext

Provides context information for use by other types in the Record Layer.

RecordVersion

Type representing a specific version within FDB.