Module fdb_rl::cursor

Provides types and traits for Record Layer cursors.

Modules

continuation 🔒

cursor 🔒

cursor_result 🔒

key_value_cursor 🔒

limit_manager 🔒

pb 🔒

Structs

CursorFilter

Cursor returned by Cursor::filter method.

CursorMap

Cursor returned by Cursor::map method.

CursorSuccess

Object of type T produced by a Cursor along with a CursorResultContinuation.

KeyValueCursor

The basic cursor for scanning ranges of the FDB database.

KeyValueCursorBuilder

A builder for KeyValueCursor. A value of KeyValueCursor can be built as shown below.

Enums

CursorError

Error that occurred when advancing the Cursor.

NoNextReason

Track reason for in-band or out-of-band cursor termination.

Traits

Continuation

Types that represent the continuation of a Cursor.

Cursor

An asynchronous iterator that supports Continuation.

Type Definitions

CursorResult

Result obtained when Cursor advances.

CursorResultContinuation

A Continuation that is present within CursorResult and returned when the Cursor advances.