Trait fdb_rl::cursor::continuation::Continuation

pub trait Continuation: Sealed + Any {
    // Required methods
    fn to_bytes(&self) -> FdbResult<Bytes>;
    fn is_begin_marker(&self) -> bool;
    fn is_end_marker(&self) -> bool;
}

Types that represent the continuation of a Cursor.

A Continuation represents the current position of a cursor and can be used to restart a cursor at a point immediately after the value returned by CursorSuccess or CursorError.

A continuation can be serialized to an opaque byte array that can be passed to a client.

Required Methods

fn to_bytes(&self) -> FdbResult<Bytes>

Serialize this continuation to a Bytes value.

fn is_begin_marker(&self) -> bool

Return whether this continuation is at the begin marker position.

Begin marker is the position in the cursor before any items are returned.

fn is_end_marker(&self) -> bool

Return whether this continuation is at the end marker position.

End marker is the position in the cursor after all the items are returned.