Type Definition fdb_rl::cursor::CursorResult

pub type CursorResult<T> = Result<CursorSuccess<T>, CursorError>;

Result obtained when Cursor advances.

It represents everything that one can learn each time a Cursor advances. It consists of two variants:

Ok(CursorSuccess<T>) is the next object of type T produced by the cursor. In addition to the next object, it also includes a CursorResultContinuation that can be used to continue the cursor after the last object is returned.

Err(CursorError<T>) represents an error condition which could be an FdbError or cursor termination due to in-band or out-of-band cursor termination (NoNextReason). In both the cases a CursorResultContinuation is included that can be used to continue the cursor.