Module fdb_rl::error

Provides error constants for use with FdbError and FdbResult.

Constants

CURSOR_INVALID_CONTINUATION

Provided continuation was invalid.

CURSOR_KEYVALUE_CURSOR_BUILDER_ERROR

Error occurred when trying to create a value of type KeyValueCursor using the build method.

RAW_RECORD_INVALID_PRIMARY_KEY_SCHEMA

The [TupleSchema] used to create a primary key is invalid.

RAW_RECORD_PRIMARY_KEY_TUPLE_SCHEMA_MISMATCH

The [Tuple] used for primary key does not match its schema.

RECORD_CONTEXT_INCARNATION_VERSION_ALREADY_SET

The incarnation version maintained within RecordContext was previously set, and cannot be updated.

RECORD_CONTEXT_INVALID_TRANSACTION_AGE

The timer maintained within RecordContext is no longer valid.

RECORD_CONTEXT_LOCAL_VERSION_OVERFLOW

The local version counter maintained within RecordContext overflowed.

SPLIT_HELPER_INVALID_PRIMARY_KEY

Primary key provided to split_helper::load, split_helper::save and split_helper::delete functions is invalid.

SPLIT_HELPER_LOAD_INVALID_RECORD_VERSION

split_helper::load function encountered an invalid record version for the primary key.

SPLIT_HELPER_LOAD_INVALID_SERIALIZED_BYTES

split_helper::load function encountered an invalid record seralized bytes for the primary key.

SPLIT_HELPER_SAVE_INVALID_SERIALIZED_BYTES_SIZE

The size of the seralized bytes provided to split_helper::save function is invalid.

SPLIT_HELPER_SCAN_LIMIT_REACHED

Scan limit reached when calling split_helper::load function.