Struct fdb_rl::key_expression::well_formed_message_descriptor::MessageDescriptorValidatorVisitor

struct MessageDescriptorValidatorVisitor {
    walked_message_descriptor: Vec<MessageDescriptor>,
}

There are two valid forms of field descriptor that we want.

cardinality: Optional, supports_presence: true, is_list: false, default_value: None, containing_oneof: Some("..."), where ... is the name of the oneof. ... can be explicitly provided when oneof keyword is used. Alternatively, when optional is specified, it is generated by the protobuf compiler.

cardinality: Repeated, supports_presence: false, is_list: true, containing_oneof: None. In proto3, there is no concept of optional repeated 1.

When cardinality: Optional, support_presence: false and is_list: false, that is a round about way of saying “required” with default value. We do not allow that. The only way to specify that a field is “required” is by using a not null key expression.

We outright reject cardinality: Required. If we see cardinality: Required, there is some serious bug, as we are checking for proto3 from file descriptor.

We also reject is_group: true, is_map: true, default_value: Some(...). There is no API from FileDescriptor type to check default_value. Therefore we need to check using FileDescriptorProto type.

Protobuf packing can vary depending on the type. So, we do not check is_packed.

Protobuf message can have no fields. Together with oneof this is very useful in building up discriminated unions (tagged union, sum types). We use this technique in key_expressions.proto. The problem with a message that has no fields is that there is no way to use Field key expression on it. While we allow message with no fields, this is something that the user must be aware of. Semantically this is equivalent to targeting a field that does not exist, so the key expression cannot be applied on the message.

We support oneof keyword. Oneof fields are just fields in the enclosing message descriptor 2 and they have containing_oneof: Some("..."), where ... specified in .proto file rather than being compiler generated. Fields that are part of oneof keyword have supports_presence: true, cardinality: Optional.

As mentioned above, for a non-repeated field, cardinality: Optional, supports_presence: true, is_list: false containing_oneof: Some("...") is the form that we want.

Additionally, within oneof, there cannot be a repeated field 3. This ensures that we won’t see an empty repeated field within oneof.

Following message type gives us the correct form.

message HelloWorld {
  optional fdb_rl.field.v1.UUID primary_key = 1;
  optional string hello = 2;
  optional string world = 3;
}

If we were to remove the optional keyword, for example:

message HelloWorld {
  fdb_rl.field.v1.UUID non_optional_primary_key = 4;
  string non_optional_string = 5;
}

Then the field non_optional_primary_key would have containing_oneof: None. The field non_optional_string would have containing_oneof: None and supports_presence: false.

We do not support protobuf map type. A map type is implemented using a repeated field with a protobuf generated message type. This generated message type consists of two fields - key and value whose types are the type parameters for the map type.

The key and value fields has support_presence: false and cardiality: Optional, which is a weird way means “required” field. The only way we allow for specifying a field is “required” is by using a key expression and not within protobuf.

Therefore, the semantics of key expression is not clear for map type. There are no good test cases from Java RecordLayer that we can use here. So, for now, we will not support map type. In the meantime, if the user needs a map type, they can use the same trick as what the protobuf compiler does (which is define a repeated field with a message and do the checking in code).

Fields
walked_message_descriptor: Vec<MessageDescriptor>
Implementations

impl MessageDescriptorValidatorVisitor

fn new() -> MessageDescriptorValidatorVisitor