Function fdb_rl::range::bytes_endpoint::build_range_continuation

pub(crate) fn build_range_continuation(
    maybe_subspace: &Option<Subspace>,
    key_range: KeyRange,
    maybe_continuation_key: Option<Bytes>,
    reverse: bool
) -> FdbResult<Range>

Build a range with an optional continuation that is not a begin marker or end marker.

Begin marker and end marker are special cases that needs to be handled seperately. That is done in KeyValueCursorBuilder::build_range method.

Note: This method along with KeyValueCursorBuilder::build_range and bytes_endpoint::build_range_bytes has extensive integration tests to verify its correctness. Exercise care when refactoring this code.