Struct fdb_rl::scan::scan_limiter::ByteScanLimiter

pub struct ByteScanLimiter {
    limiter_mode: LimiterMode,
    limit: usize,
    bytes_scanned: Arc<AtomicUsize>,
}

Track the number of bytes scanned up to some limit, after which scans should not be allowed.

Fields
limiter_mode: LimiterModelimit: usizebytes_scanned: Arc<AtomicUsize>
Implementations

impl ByteScanLimiter

pub fn enforcing(limit: usize) -> ByteScanLimiter

Create a value of ByteScanLimiter that enforces a maximum number of bytes that can be processed in a single scan.

pub fn tracking() -> ByteScanLimiter

Create a value of ByteScanLimiter that tracks the number of bytes that has been scanned, but does not impose a limit.

pub fn untracked() -> ByteScanLimiter

Create a value of ByteScanLimiter that neither enforces nor tracks the number of bytes scanned.

pub fn get_limit(&self) -> usize

Get the byte scan limit.

pub fn get_scanned_bytes(&self) -> usize

Returns the number of bytes that have been scanned thus far.

pub fn is_enforcing(&self) -> bool

Return whether or not this limiter is actually enforcing the limit.

pub(crate) fn try_keyvalue_scan(&self) -> bool

Return true if keyvalue can be read otherwise return false.

Note: In Java RecordLayer, this method is called hasBytesRemaining. We are renaming it so that the API is consistent with KeyValueScanLimiter and ByteScanLimiter.

Used by LimitManager::try_keyvalue_scan (Linking does not work!).

pub(crate) fn register_scanned_bytes(&self, bytes: usize)

Increment the number of bytes scanned by the given number of bytes.

Used by LimitManager::register_scanned_bytes (Linking does not work!).