Function fdb_rl::split_helper::delete_unchecked

unsafe fn delete_unchecked<Tr>(
    tr: &Tr,
    maybe_subspace: &Option<Subspace>,
    primary_key: &Tuple
) -> FdbResult<()>where
    Tr: Transaction,

Delete the serialized representation of a record without any safety checks.

Note

You will never want to use this function. Any mistake with the primary_key can seriously damage the database, as it will issue a Transaction::clear_range without any checks.

The only place where this is useful is in the delete and save methods, where we have already checked the validity of primary_key.