Module fdb_rl::test::raw_record

Structs

RawRecord

A wrapper around all information that can be determined about a record before serializing and deserializing it.

RawRecordPrimaryKey

Represents a RawRecord’s primary key which is a Tuple.

RawRecordPrimaryKeySchema

Represents the schema for a RawRecordPrimaryKey.

Constants

RAW_RECORD_CURSOR_BUILDER_ERROR

Error occurred when trying to create a value of type RawRecordCursor using the build method.

RAW_RECORD_CURSOR_NEXT_ERROR

Error occured when parsing the key or value

Functions

raw_record_cursor_builder_build

Expose RawRecordCursorBuilder::build method.

raw_record_from

Expose RawRecord::from method.