Module fdb_rl::test::split_helper

Constants

SPLIT_HELPER_INVALID_PRIMARY_KEY

Primary key provided to split_helper::load, split_helper::save and split_helper::delete functions is invalid.

SPLIT_HELPER_LOAD_INVALID_RECORD_HEADER

split_helper::load function encountered an invalid record header for the primary key.

SPLIT_HELPER_LOAD_INVALID_SERIALIZED_BYTES

split_helper::load function encountered an invalid record seralized bytes for the primary key.

SPLIT_HELPER_SAVE_INVALID_SERIALIZED_BYTES_SIZE

The size of the seralized bytes provided to split_helper::save function is invalid.

SPLIT_HELPER_SCAN_LIMIT_REACHED

Scan limit reached when calling split_helper::load function.

Functions

delete

Delete the serialized representation of a record.

load

Load serialized byte array that may be split among several keys.

save

Save serialized representation using multiple keys if necessary.