Function fdb_rl::test::split_helper::load

pub async fn load<Tr>(
    tr: &Tr,
    maybe_scan_limiter: &Option<ScanLimiter>,
    maybe_subspace: &Option<Subspace>,
    primary_key: &Tuple
) -> FdbResult<Option<(RecordVersion, Bytes)>>where
    Tr: ReadTransaction,

Load serialized byte array that may be split among several keys.

When a value of Ok(None) is returned, it means that there is no serialized byte array associated with the primary_key. If there is a serialized byte array associated with the primary_key, then we would return Ok(Some((record_version, seralized_bytes))) value. Otherwise, an Err value is returned.

Warning: This function is not meant to be public. We need to make this function public to support integration tests. Do not use this function in your code.