Module fdb_rl::key_expression::pb

Protobuf types.

Structs

ConcatInternalV1 🔒

Protobuf message fdb_rl.key_expression.v1.Concat uses protos that contain a Required field. So, we need to define this type.

FieldFanTypeInternalV1 🔒

Protobuf message fdb_rl.key_expression.v1.Field.FieldFanType contains a Required field. So, we need to define this type.

FieldInternalV1 🔒

Protobuf message fdb_rl.key_expression.v1.Field contains a Required field. So, we need to define this type.

FieldNullInterpretationInternalV1 🔒

Protobuf message fdb_rl.key_expression.v1.Field.FieldNullInterpretation contains a Required field. So, we need to define this type.

KeyExpressionInternalV1 🔒

Protobuf message fdb_rl.key_expression.v1.KeyExpression contains a Required field. So, we need to define this type.

NestInternalV1 🔒

Protobuf message fdb_rl.key_expression.v1.Nest contains a Required field. So, we need to define this type.

ProtoConcatV1 🔒

Protobuf generated types renamed to prepend Proto and append version. TODO

ProtoConcatenateV1 🔒

Protobuf generated types renamed to prepend Proto and append version (and add Enum suffix). Convert a repeated field into a single list.

ProtoDistinctV1 🔒

Protobuf generated types renamed to prepend Proto and append version (and add Enum suffix). Missing value (NULL) are allowed multiple times in unique index (PostgreSQL NULLS DISTINCT, Java RecordLayer NOT_UNIQUE). This is the default for PostgreSQL and Java RecordLayer.

ProtoFanOutV1 🔒

Protobuf generated types renamed to prepend Proto and append version (and add Enum suffix). Create an index value for each value of the field.

ProtoFieldFanTypeV1 🔒

Protobuf generated types renamed to prepend Proto and append version (and add Enum suffix). How to handle repeated field.

ProtoFieldNullInterpretationV1 🔒

Protobuf generated types renamed to prepend Proto and append version (and add Enum suffix). Field presence is the notion of whether a protobuf field has a value [1]. As of release 3.15, in proto3, the default is explicit presence, which is managed by optional label.

ProtoFieldV1 🔒

Protobuf generated types renamed to prepend Proto and append version. Protobuf message field is specified here [1].

ProtoKeyExpressionV1 🔒

Protobuf generated types renamed to prepend Proto and append version. TODO

ProtoNestV1 🔒

Protobuf generated types renamed to prepend Proto and append version. TODO

ProtoNotDistinctV1 🔒

Protobuf generated types renamed to prepend Proto and append version (and add Enum suffix). Missing value (NULL) cannot be repeated in a unique index. This is very restrictive as a column can have only one NULL value. (PostgreSQL NULLS NOT DISTINCT, Java RecordLayer UNIQUE).

ProtoNotNullV1 🔒

Protobuf generated types renamed to prepend Proto and append version (and add Enum suffix). Field is not NULL. As mentioned in the above table, in proto3 this is only possible for a singular numeric, enum, bytes, string, bool type.

Enums

FieldLabelInternalEnumV1 🔒

Protobuf message fdb_rl.key_expression.v1.Field.field_label enum indirectly contains a Required field (via FieldNullInterpretation and FieldFanType. So, we need to define this type.

KeyExpressionInternalEnumV1 🔒

Protobuf oneof fdb_rl.key_expression.v1.KeyExpression.key_expression uses protos that contain a Required field. So, we need to define this type.

ProtoFieldFanTypeEnumV1 🔒

Protobuf generated types renamed to prepend Proto and append version (and add Enum suffix). Required.

ProtoFieldLabelEnumV1 🔒

Protobuf generated types renamed to prepend Proto and append version (and add Enum suffix). Required.

ProtoFieldNullInterpretationEnumV1 🔒

Protobuf generated types renamed to prepend Proto and append version (and add Enum suffix). Required.

ProtoKeyExpressionEnumV1 🔒

Protobuf generated types renamed to prepend Proto and append version (and add Enum suffix). Required.