Module fdb_rl::range

Provides types for Record Layer range.

Record Layer range module provides a higher level of abstraction for working FDB range and cursors.

Please consider using TupleRange and only if the APIs are not suitable, then use KeyRange.

Modules

bytes_endpoint 🔒

endpoint 🔒

key_range 🔒

tuple_range 🔒

Structs

KeyRange

A range specified by two Key endpoints.

TupleRange

A range within an optional subspace specified by two Tuple endpoints.

Enums

HighEndpoint

High endpoint of a range.

LowEndpoint

Low endpoint of a range.

Type Aliases

KeyHighEndpoint

High endpoint of a Key.

KeyLowEndpoint

Low endpoint of a Key.

TupleHighEndpoint

High endpoint of a TupleRange.

TupleLowEndpoint

Low endpoint of a TupleRange.